Engine/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-autoslug==1.9.9
django-cors-headers==4.1.0
django-elasticsearch-dsl==7.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
twilio==8.3.0
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
